MyApps-13: Move book selection UseCase to submodules.
