+++ app/src/main/java/xyz/torquato/myapps/ui/component/web/WebViewModel.kt
@@ -9,11 +9,11 @@ import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.stateIn
 import xyz.torquato.myapps.domain.web.SetQueryUseCase
-import xyz.torquato.myapps.domain.web.SetSelectedBookUseCase
 import xyz.torquato.myapps.domain.web.content.GetBooksUseCase
 import xyz.torquato.myapps.domain.web.pagging.GetNextPageUseCase
 import xyz.torquato.myapps.ui.component.web.model.BookMenuMapper.toUiState
 import xyz.torquato.myapps.ui.component.web.model.BookMenuUiState
+import xyz.torquato.myaps.domain.impl.web.content.SetSelectedBookUseCase
 import javax.inject.Inject
 
 @HiltViewModel
