+++ domain/impl/src/main/java/xyz/torquato/myaps/domain/impl/web/content/SetSelectedBookUseCase.kt
@@ -1,4 +1,4 @@
-package xyz.torquato.myapps.domain.web
+package xyz.torquato.myaps.domain.impl.web.content
 
 import xyz.torquato.myaps.domain.api.web.ILocalWebRepository
